refactor: domain type 추출
