+++ src/domain/types/card.ts
@@ -1,0 +1,14 @@
+type CardInformation = {
+  cardNumber: CardNumber;
+  expirationDate: ExpirationDate;
+  owner?: string;
+};
+
+type CardNumber = [string, string, string, string];
+
+type ExpirationDate = {
+  year: string;
+  month: string;
+};
+
+export type { CardInformation, CardNumber, ExpirationDate };
